pyproject.toml
[project]
name = "reya-python-sdk-ccxt-wrapper"
version = "2.0.3"
description = "CCXT Wrapper and SDK for interacting with Reya Labs APIs"
authors = [
    {name = "Reya Labs and marcelkb"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "web3>=7.12.1,<8.0.0",
    "eth-account>=0.13.7,<1.0.0",
    "eth-abi>=5.2.0,<6.0.0",
    "typing-extensions>=4.14.1,<5.0.0",
    "python-dotenv>=1.1.1,<2.0.0",
    "requests>=2.32.4,<3.0.0",
    "httpx>=0.28.1,<1.0.0",
    "websocket-client>=1.8.0,<2.0.0",
    "tomli>=2.0.1,<3.0.0",
    "types-requests>=2.31.0,<2.32.0",
    "python-dateutil>=2.8.0,<3.0.0",
    "aiohttp-retry>=2.8.3,<3.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<=1.1.0",
    "black>=24.7.0,<25.0.0",
    "isort>=5.12.0,<6.0.0",
    "flake8>=7.0.0,<8.0.0",
    "pytest-recording>=0.13.2,<1.0.0",
    "mypy>=1.14.1,<2.0.0",
    "mypy-extensions>=1.0.0,<2.0.0",
    "pre-commit>=4.1.0,<5.0.0",
    "safety>=3.2.14,<4.0.0",
    "coverage>=7.6.10,<8.0.0",
    "pytest-cov>=6.0.0,<7.0.0",
    "vcrpy>=7.0.0,<8.0.0",
    "types-requests>=2.31.0,<2.32.0",
    "lz4>=4.3,<5.0"
]

[tool.poetry]
packages = [
    {include = "examples"},
    {include = "sdk"},
    {include = "ccxt_wrapper"}
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
extend-exclude = "sdk/async_api"

[tool.isort]
# https://github.com/timothycrosley/isort/
py_version = 310
line_length = 120
extend_skip_glob = ["sdk/async_api/*"]

known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true

[tool.mypy]
# https://mypy.readthedocs.io/en/latest/config_file.html#using-a-pyproject-toml-file
python_version = "3.10"
pretty = true
show_traceback = true
color_output = true
exclude = ["sdk/async_api/"]

allow_redefinition = false
check_untyped_defs = true
disallow_any_generics = false
disallow_incomplete_defs = false
ignore_missing_imports = true
implicit_reexport = true
no_implicit_optional = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

# Ignore specific modules with type issues
[[tool.mypy.overrides]]
module = "sdk.open_api.*"
ignore_errors = true

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 120
max-module-lines=2000
ignore-paths = ["sdk/async_api"]
disable = """
    consider-using-in,
    duplicate-code,
    fixme,
    import-error,
    inconsistent-return-statements,
    invalid-name,
    logging-fstring-interpolation,
    line-too-long,
    missing-class-docstring,
    missing-function-docstring,
    missing-module-docstring,
    no-else-raise,
    no-else-return,
    raise-missing-from,
    redefined-builtin,
    self-assigning-variable,
    simplifiable-if-expression,
    too-few-public-methods,
    too-many-arguments,
    too-many-branches,
    too-many-positional-arguments,
    too-many-public-methods,
    too-many-return-statements,
    too-many-instance-attributes
"""

[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
norecursedirs =["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]
